requirements.txt
playwright==1.44.0
supabase
python-dotenv
schedule
python-telegram-bot